chore: moved mui css initialization to theme wrapper to align test renders
